Add Counter.
This includes a new page accessible through the path
"/counter/:name" in the router.

The page uses a new component, CardDialog, that is
also used directly in CardGrid to open a dialog
and allow to change the counter directly.

The value of the counter for each name is stored
using localStorage.