Add a tree view.
This includes a new page accessible through the path
"/organisations" in the router.

One can access this page from a new button in the
frontpage (landing page).

The data is read from data/stubs/OrganisationUnits.json.
A future version will read it directly from
https://dev.eyeseetea.com/play .

When selecting a unit, a panel to the right gets filled
with information related to it. If one changes any of
those fields, a couple of buttons appear to let the user
save the changes (a stub - the saving is not implemented
yet).